+++ tasks/generate.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-# Copyright (C) 2017 Damon Lynch <damonlynch@gmail.com>
+# Copyright (C) 2017-2018 Damon Lynch <damonlynch@gmail.com>
 
 # This file is part of Rapid Photo Downloader.
 #
@@ -19,11 +19,11 @@
 # see <http://www.gnu.org/licenses/>.
 
 """
-Create a data blob of locale mo files and insert it into install.py
+Create a data blob of locale mo files and insert it into install.py and upgrade.py
 """
 
 __author__ = 'Damon Lynch'
-__copyright__ = "Copyright 2017, Damon Lynch"
+__copyright__ = "Copyright 2017-2018, Damon Lynch"
 
 
 from glob import glob